requirements.txt
Ghost.py
requests
jinja2
bs4